+++ lib/ts/framework/fastify/framework.ts
@@ -13,12 +13,12 @@
  * under the License.
  */
 
-import type {
-    FastifyInstance,
+import {
     FastifyRequest as OriginalFastifyRequest,
     FastifyReply,
+    FastifyInstance,
     FastifyPluginCallback,
-} from "fastify";
+} from "./types";
 import type { HTTPMethod } from "../../types";
 import { getFromObjectCaseInsensitive, makeDefaultUserContextFromAPI, normaliseHttpMethod } from "../../utils";
 import { BaseRequest } from "../request";
@@ -133,9 +133,9 @@ export class FastifyResponse extends BaseResponse {
     ) => {
         let serialisedCookie = serializeCookieValue(key, value, domain, secure, httpOnly, expires, path, sameSite);
 
-        let oldHeaders: string | string[] | undefined = this.response.getHeader(COOKIE_HEADER);
+        let oldHeaders: string | string[] | undefined | number = this.response.getHeader(COOKIE_HEADER);
         if (oldHeaders === undefined) oldHeaders = [];
-        else if (!((oldHeaders as any) instanceof Array)) oldHeaders = [oldHeaders];
+        else if (!((oldHeaders as any) instanceof Array)) oldHeaders = [oldHeaders.toString()];
 
         this.response.removeHeader(COOKIE_HEADER);
         this.response.header(COOKIE_HEADER, [
@@ -164,7 +164,7 @@ export class FastifyResponse extends BaseResponse {
     };
 }
 
-function plugin(fastify: FastifyInstance, _: any, done: Function) {
+function plugin(fastify: FastifyInstance, _: unknown, done: () => void) {
     fastify.addHook("preHandler", async (req: OriginalFastifyRequest, reply: FastifyReply) => {
         let supertokens = SuperTokens.getInstanceOrThrowError();
         let request = new FastifyRequest(req);
@@ -185,7 +185,7 @@ export type SessionRequest<TRequest extends OriginalFastifyRequest = OriginalFas
     session?: SessionContainerInterface;
 };
 
-export interface FasitfyFramework extends Framework {
+export interface FastifyFramework extends Framework {
     plugin: FastifyPluginCallback;
     errorHandler: () => (err: any, req: OriginalFastifyRequest, res: FastifyReply) => Promise<void>;
 }
@@ -200,7 +200,7 @@ export const errorHandler = () => {
     };
 };
 
-export const FastifyWrapper: FasitfyFramework = {
+export const FastifyWrapper: FastifyFramework = {
     plugin,
     errorHandler,
     wrapRequest: (unwrapped) => {