+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added a way to run CI on unmerged PRs
 -   Added support for FDIs: 3.1 and 4.0. Required by: auth-react >=0.49.0 and web-js>=0.15.0
 -   The `networkInterceptor` now also gets a new `params` prop in the request config.
+-   Replicates `fastify` types based on requirement for the SDK instead of using the original module.
 
 ### Breaking change